refactor(profile): replace param `user-id` -> `id`
