+++ app/api/hooks/use-recently-played-query.tsx
@@ -4,14 +4,14 @@ import { useParams } from 'next/navigation'
 import { RECENTLY_PLAYED } from '../constants'
 import { getRecentlyPlayed } from '../fetchers'
 
-import { USER_ID } from '@app/constants'
+import { ID } from '@app/constants'
 import { useToken } from '@app/hooks/use-token'
 
 export const useRecentlyPlayedQuery = () => {
   const token = useToken()
   const params = useParams()
 
-  const userId = params[USER_ID].toString()
+  const userId = params[ID].toString()
 
   return useQuery({
     queryKey: [RECENTLY_PLAYED],
