+++ app/api/hooks/use-toggle-playback-state.tsx
@@ -2,14 +2,14 @@ import { useParams } from 'next/navigation'
 
 import { putPlayerPause, putPlayerResume } from '../fetchers'
 
-import { USER_ID } from '@app/constants'
+import { ID } from '@app/constants'
 import { useToken } from '@app/hooks/use-token'
 
 export const useTogglePlaybackStateQuery = () => {
   const token = useToken()
   const params = useParams()
 
-  const userId = params[USER_ID].toString()
+  const userId = params[ID].toString()
 
   function toggle(isPlaying = false) {
     return isPlaying
