+++ app/profile/[id]/@analysis/loading.tsx
