+++ app/profile/[id]/@top/@albums/loading.tsx
