+++ app/profile/[id]/error.tsx
