+++ app/profile/[id]/history/loading.tsx
