+++ app/profile/[id]/history/page.tsx
@@ -2,13 +2,13 @@ import { redirect } from 'next/navigation'
 
 import { validateId } from '@app/utils/validate-id'
 import type { ProfilePageProps } from '@app/profile/types'
-import { USER_ID } from '@app/constants'
+import { ID } from '@app/constants'
 import { getHistory } from '@app/api/fetchers'
 import { HistorySection } from '@app/profile/sections'
 import { getServerToken } from '@app/auth/utils'
 
 export default async function ProfileHistoryPage({ params }: ProfilePageProps) {
-  const userId = validateId(params[USER_ID])
+  const userId = validateId(params[ID])
   const limit = 20
 
   const token = await getServerToken()
